Add application round time slots to admin panel
Also updates libraries and adds new "django-subforms" library for
better admin interface for saving them.
