+++ pyproject.toml
@@ -54,6 +54,7 @@ redis = "^5.0.0"
 sentry-sdk = "^1.29.2"
 social-auth-app-django = "^5.2.0"
 whitenoise = "^6.5.0"
+django-subforms = "^0.2.0"
 python-dateutil = "^2.8.2"
 
 [tool.poetry.dependencies.uWSGI]
