+++ reservation_units/admin/day_part.py
@@ -1,0 +1,8 @@
+from django.contrib import admin
+
+from reservation_units.models import DayPart
+
+
+@admin.register(DayPart)
+class DayPartAdmin(admin.ModelAdmin):
+    model = DayPart
