+++ reservation_units/admin/keyword.py
@@ -1,0 +1,18 @@
+from django.contrib import admin
+
+from reservation_units.models import Keyword, KeywordCategory, KeywordGroup
+
+
+@admin.register(KeywordCategory)
+class KeywordCategoryAdmin(admin.ModelAdmin):
+    model = KeywordCategory
+
+
+@admin.register(KeywordGroup)
+class KeywordGroupAdmin(admin.ModelAdmin):
+    model = KeywordGroup
+
+
+@admin.register(Keyword)
+class KeywordAdmin(admin.ModelAdmin):
+    model = Keyword
