+++ reservation_units/admin/period.py
@@ -1,0 +1,8 @@
+from django.contrib import admin
+
+from reservation_units.models import Period
+
+
+@admin.register(Period)
+class PeriodAdmin(admin.ModelAdmin):
+    model = Period
