+++ reservation_units/admin/qualifier.py
@@ -1,0 +1,8 @@
+from django.contrib import admin
+
+from reservation_units.models import Qualifier
+
+
+@admin.register(Qualifier)
+class QualifierAdmin(admin.ModelAdmin):
+    model = Qualifier
