+++ reservation_units/admin/reservation_unit_cancellation_rule.py
@@ -1,0 +1,8 @@
+from django.contrib import admin
+
+from reservation_units.models import ReservationUnitCancellationRule
+
+
+@admin.register(ReservationUnitCancellationRule)
+class ReservationUnitCancellationRuleAdmin(admin.ModelAdmin):
+    model = ReservationUnitCancellationRule
