+++ reservation_units/admin/reservation_unit_image.py
@@ -1,0 +1,8 @@
+from django.contrib import admin
+
+from reservation_units.models import ReservationUnitImage
+
+
+@admin.register(ReservationUnitImage)
+class ReservationUnitImageAdmin(admin.ModelAdmin):
+    model = ReservationUnitImage
