+++ reservation_units/admin/reservation_unit_pricing.py
@@ -1,0 +1,8 @@
+from django.contrib import admin
+
+from reservation_units.models import ReservationUnitPricing
+
+
+@admin.register(ReservationUnitPricing)
+class ReservationUnitPricingAdmin(admin.ModelAdmin):
+    model = ReservationUnitPricing
