+++ reservation_units/admin/reservation_unit_type.py
@@ -1,0 +1,9 @@
+from adminsortable2.admin import SortableAdminMixin
+from django.contrib import admin
+
+from reservation_units.models import ReservationUnitType
+
+
+@admin.register(ReservationUnitType)
+class ReservationUnitTypeAdmin(SortableAdminMixin, admin.ModelAdmin):
+    model = ReservationUnitType
