+++ reservation_units/admin/tax_percentage.py
@@ -1,0 +1,8 @@
+from django.contrib import admin
+
+from reservation_units.models import TaxPercentage
+
+
+@admin.register(TaxPercentage)
+class TaxPercentageAdmin(admin.ModelAdmin):
+    model = TaxPercentage
